@@ -1,4 +1,7 @@
-import ReactDOM from 'react-dom';
-import App from './App';
+import React from 'react';
+import { createRoot } from 'react-dom/client';
+import App from './components/app';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+const container = document.getElementById('root');
+const root = createRoot(container!);
+root.render(<App />);
