@@ -1,7 +1,14 @@
-export type FetchCatsRequest = { type: string }; // TODO : Update this type !
-export type FetchCatsCommit = { type: string }; // TODO : Update this type !
-export type FetchCatsRollback = { type: string }; // TODO : Update this type !
+import { Decrement, FetchCatsCommit, FetchCatsRequest, FetchCatsRollback, Increment } from './types/actions.type';
 
-export const fetchCatsRequest = () => {}; // TODO : Update this value !
-export const fetchCatsCommit = (payload: unknown) => {}; // TODO : Update this value !
-export const fetchCatsRollback = (error: Error) => {}; // TODO : Update this value !
+export const increment = (): Increment => ({ type: 'INCREMENT' });
+export const decrement = (): Decrement => ({ type: 'DECREMENT' });
+
+export const fetchCatsRequest = (): FetchCatsRequest => ({
+  type: 'FETCH_CATS_REQUEST',
+  method: 'GET',
+  path: 'Update the path',
+}); // TODO : Update this value !
+
+export const fetchCatsCommit = (payload: unknown): FetchCatsCommit => ({ type: 'FETCH_CATS_COMMIT', payload });
+
+export const fetchCatsRollback = (error: Error): FetchCatsRollback => ({ type: 'FETCH_CATS_ROLLBACK', error });
